Move local/remote commit tx construction
